# Backend/docs/swagger.yaml
basePath: /api/v2/
definitions:
...
  controller.ErrorResponse:
    description: Error response
    properties:
      error:
        description: '@Description Error message'
        example: error message
        type: string
    type: object
...
  controller.WalletResponse:
    description: Wallet information response
    properties:
      balance:
        description: '@Description Current balance in the wallet'
        example: 1000.5
        type: number
      created_at:
        description: '@Description When the wallet was created'
        example: "2024-05-22T09:00:00Z"
        type: string
      currency:
        description: '@Description Currency code'
        example: ETB
        type: string
      description:
        description: '@Description Optional description of the wallet'
        example: Main merchant wallet
        type: string
      id:
        description: '@Description Unique identifier for the wallet'
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      is_active:
        description: '@Description Whether the wallet is active'
        example: true
        type: boolean
      last_sync_at:
        description: '@Description When the wallet was last synchronized'
        example: "2024-05-22T09:00:00Z"
        type: string
      merchant_id:
        description: '@Description ID of the merchant who owns this wallet'
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      updated_at:
        description: '@Description When the wallet was last updated'
        example: "2024-05-22T09:00:00Z"
        type: string
    type: object
...
  entity.MerchantCommission:
    properties:
      commission_active:
        type: boolean
      commission_cent:
        type: number
      commission_percent:
        type: number
      merchant_id:
        type: string
    type: object
...
  entity.QRMerchantPaymentRequest:
    description: QR Payment request details for processing a payment from checkout
    properties:
      amount:
        description: |-
          Amount to be paid
          @Example 1000.50
        example: 1000.5
        type: number
      currency:
        description: |-
          Optional: Three-letter currency code (defaults to ETB)
          @Example ETB
        example: ETB
        type: string
      description:
        description: |-
          Optional: Description of the payment
          @Example Payment for order #123
        example: 'Payment for order #123'
        type: string
      medium:
        allOf:
        - $ref: '#/definitions/github_com_socialpay_socialpay_src_pkg_transaction_core_entity.TransactionMedium'
        description: |-
          Payment medium/provider to use
          @Example MPESA
        example: MPESA
      merchant_id:
        description: |-
          Merchant ID
          @Example 123e4567-e89b-12d3-a456-426614174000
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      phone_number:
        description: |-
          Optional: Phone number to be paid
          @Example 251911111111
        example: "251911111111"
        type: string
      reference:
        description: |-
          Optional: Client-provided reference
          @Example ORDER123456
        example: ORDER123456
        type: string
    type: object
...
  gin.DeleteMerchantsRequest:
    properties:
      ids:
        items:
          type: string
        type: array
    type: object
...
  github_com_socialpay_socialpay_src_pkg_socialpayapi_adapter_controller_gin.ErrorResponse:
    description: Error response with a message
    properties:
      message:
        example: Invalid request parameters
        type: string
      success:
        type: boolean
    type: object
  github_com_socialpay_socialpay_src_pkg_socialpayapi_core_entity.QRPaymentResponse:
    description: Response containing QR payment processing results
    properties:
      socialpay_transaction_id:
        description: SocialPay transaction ID
        example: "1234567890"
        type: string
      message:
        description: Human-readable message about the operation
        example: Payment initiated successfully
        type: string
      payment_url:
        description: URL to redirect the user for payment (if applicable)
        example: https://pay.example.com/checkout
        type: string
      reference_id:
        description: Unique payment reference
        example: PAY123456789
        type: string
      status:
        description: Status of the payment
        example: PENDING
        type: string
      success:
        description: Whether the operation was successful
        example: true
        type: boolean
    type: object
...
  github_com_socialpay_socialpay_src_pkg_team_member_adapter_controller_gin.ApiError:
    properties:
      message:
        type: string
      type:
        type: string
    type: object
  github_com_socialpay_socialpay_src_pkg_team_member_adapter_controller_gin.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_socialpay_socialpay_src_pkg_team_member_adapter_controller_gin.ApiError'
      success:
        type: boolean
    type: object
  github_com_socialpay_socialpay_src_pkg_team_member_adapter_controller_gin.SuccessResponse:
    properties:
      data: {}
      success:
        type: boolean
    type: object
  github_com_socialpay_socialpay_src_pkg_transaction_core_entity.TransactionMedium:
    description: Payment medium or provider (e.g., MPESA, TELEBIRR)
    enum:
    - SOCIALPAY
    - CYBERSOURCE
    - ETHSWITCH
    - MPESA
    - TELEBIRR
    - CBE
    - AWASH
    - KACHA
    type: string
    x-enum-varnames:
    - SOCIALPAY
    - CYBERSOURCE
    - ETHSWITCH
    - MPESA
    - TELEBIRR
    - CBE
    - AWASH
    - KACHA
...
  github_com_socialpay_socialpay_src_pkg_v2_merchant_adapter_controller_gin.ApiError:
    properties:
      message:
        type: string
      type:
        type: string
    type: object
  github_com_socialpay_socialpay_src_pkg_v2_merchant_adapter_controller_gin.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_socialpay_socialpay_src_pkg_v2_merchant_adapter_controller_gin.ApiError'
      success:
        type: boolean
    type: object
  github_com_socialpay_socialpay_src_pkg_v2_merchant_adapter_controller_gin.SuccessResponse:
    properties:
      data: {}
      success:
        type: boolean
    type: object
...
  pagination.PaginationInfo:
    properties:
      has_next_page:
        type: boolean
      has_pervious:
        type: boolean
      page:
        minimum: 1
        type: integer
      page_size:
        maximum: 20
        minimum: 1
        type: integer
      total_items:
        type: integer
      total_page:
        type: integer
    required:
    - page
    - page_size
    type: object
...
  response.PaginatedResponse:
    properties:
      data:
        type: object
      pagination:
        $ref: '#/definitions/pagination.PaginationInfo'
      success:
        type: boolean
    type: object
...
host: 196.190.251.194:8082
info:
  contact:
    email: support@socialpay.com
    name: API Support
    url: http://www.swagger.io/support
  description: SocialPay API V2 documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SocialPay API V2
  version: "2.0"
paths:
  /admin/all:
    get:
      description: Get list of admin team members
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_socialpay_socialpay_src_pkg_team_member_adapter_controller_gin.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_socialpay_socialpay_src_pkg_team_member_adapter_controller_gin.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_socialpay_socialpay_src_pkg_team_member_adapter_controller_gin.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_socialpay_socialpay_src_pkg_team_member_adapter_controller_gin.ErrorResponse'
      summary: Get admin team members
      tags:
      - Team-member-management
...
  /admin/commissions/merchant/{merchantID}:
    get:
      consumes:
      - application/json
      description: Retrieves the commission settings for a specific merchant
      parameters:
      - description: Merchant ID
        format: uuid
        in: path
        name: merchantID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'commission: entity.MerchantCommission'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: invalid merchant ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: forbidden'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: error message'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get merchant commission settings
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Updates the commission settings for a specific merchant
      parameters:
      - description: Merchant ID
        format: uuid
        in: path
        name: merchantID
        required: true
        type: string
      - description: Merchant Commission Settings
        in: body
        name: commission
        required: true
        schema:
          $ref: '#/definitions/entity.MerchantCommission'
      produces:
      - application/json
      responses:
        "200":
          description: 'commission: entity.MerchantCommission'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: invalid request'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: forbidden'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: error message'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update merchant commission settings
      tags:
      - admin
...
  /admin/wallet:
    get:
      consumes:
      - application/json
      description: Retrieves the admin wallet information including total amount
      produces:
      - application/json
      responses:
        "200":
          description: 'wallet: entity.MerchantWallet, total_amount: float64'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: error message'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: error message'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get admin wallet
      tags:
      - admin
  /all:
    get:
      description: Get list of team members
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_socialpay_socialpay_src_pkg_team_member_adapter_controller_gin.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_socialpay_socialpay_src_pkg_team_member_adapter_controller_gin.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_socialpay_socialpay_src_pkg_team_member_adapter_controller_gin.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_socialpay_socialpay_src_pkg_team_member_adapter_controller_gin.ErrorResponse'
      summary: Get team members
      tags:
      - Team-member-management
...
  /merchants/:id:
    get:
      description: Get a merchant by its ID
      parameters:
      - description: Merchant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_socialpay_socialpay_src_pkg_v2_merchant_adapter_controller_gin.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_socialpay_socialpay_src_pkg_v2_merchant_adapter_controller_gin.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_socialpay_socialpay_src_pkg_v2_merchant_adapter_controller_gin.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_socialpay_socialpay_src_pkg_v2_merchant_adapter_controller_gin.ErrorResponse'
      summary: Get merchant by ID
      tags:
      - v2-merchants
...
  /merchants/admin/delete/all:
    delete:
      consumes:
      - application/json
      description: Admin deletes list of merchants
      parameters:
      - description: DeleteMerchants details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/gin.DeleteMerchantsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Merchantd deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Merchant not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Delete merchants by admin
      tags:
      - v2-merchants
...
  /merchants/all:
    get:
      description: Get a list of merchants
      parameters:
      - description: Merchant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_socialpay_socialpay_src_pkg_v2_merchant_adapter_controller_gin.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_socialpay_socialpay_src_pkg_v2_merchant_adapter_controller_gin.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_socialpay_socialpay_src_pkg_v2_merchant_adapter_controller_gin.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_socialpay_socialpay_src_pkg_v2_merchant_adapter_controller_gin.ErrorResponse'
      summary: Get merchants
      tags:
      - v2-merchants
...
  /qr/payment/merchant:
    post:
      consumes:
      - application/json
      description: Process a QR payment transaction from checkout page
      parameters:
      - description: QR Payment request details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.QRMerchantPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_socialpay_socialpay_src_pkg_socialpayapi_core_entity.QRPaymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_socialpay_socialpay_src_pkg_socialpayapi_adapter_controller_gin.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_socialpay_socialpay_src_pkg_socialpayapi_adapter_controller_gin.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_socialpay_socialpay_src_pkg_socialpayapi_adapter_controller_gin.ErrorResponse'
      summary: Process a QR payment for merchant
      tags:
      - QR-Payments
...
  /wallet:
    get:
      consumes:
      - application/json
      description: Get the wallet information for the authenticated merchant
      produces:
      - application/json
      responses:
        "200":
          description: Wallet information
          schema:
            $ref: '#/definitions/controller.WalletResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Wallet Not Found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - BearerAuth: []
      - MerchantID: []
      summary: Get merchant wallet
      tags:
      - wallet
...
  /webhooks/callback/merchant:
    get:
      consumes:
      - application/json
      description: Retrieves all callback logs for the merchant associated with the
        API key. The merchant ID is automatically determined from the API key context.
      parameters:
      - description: 'Page number (min: 1)'
        in: query
        name: page
        required: true
        type: integer
      - description: 'Number of items per page (min: 1, max: 100)'
        in: query
        name: page_size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PaginatedResponse'
        "400":
          description: 'error: invalid merchant ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: internal server error'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get callback logs for authenticated merchant
      tags:
      - webhooks
security:
- BearerAuth: []
securityDefinitions:
  ApiKeyAuth:
    description: API key authentication for API key management endpoints
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'Bearer token authentication for protected endpoints. Use format:
      "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
  MerchantID:
    description: Merchant ID authentication for merchant endpoints
    in: header
    name: X-MERCHANT-ID
    type: apiKey
swagger: "2.0"